feat: UX 증진을 위한 layout 추가
