+++ mirror-soul/app/signup/_layout.tsx
@@ -1,0 +1,55 @@
+import React from 'react';
+import { SafeAreaView, StatusBar, StyleSheet, Platform, View } from 'react-native';
+import { Slot, usePathname } from 'expo-router';
+import { Colors } from '@/src/constants/theme';
+import SignupBackground from '@/src/components/signup/Account/SignupBackground';
+import OnboardingSteps from '@/src/components/signup/common/OnboardingSteps';
+
+export default function SignupLayout() {
+  const pathname = usePathname();
+
+  // 현재 경로에 따라 Step 결정
+  // /signup or /signup/  -> Step 1
+  // /signup/profile      -> Step 2
+  const getCurrentStep = () => {
+    if (pathname === '/signup' || pathname === '/signup/') return 1;
+    if (pathname.includes('/profile')) return 2;
+    return 1;
+  };
+
+  return (
+    <SafeAreaView style={styles.safeArea}>
+      <StatusBar barStyle="light-content" backgroundColor="#000" />
+      <SignupBackground />
+
+      <View style={styles.container}>
+        <View style={styles.stepsWrapper}>
+          <OnboardingSteps currentStep={getCurrentStep()} />
+        </View>
+
+        {/* 하위 페이지(index, profile 등)가 렌더링될 영역 */}
+        <Slot />
+      </View>
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({
+  safeArea: {
+    flex: 1,
+    backgroundColor: Colors.primary.soulBlack,
+    paddingTop: Platform.OS === 'android' ? StatusBar.currentHeight : 0,
+  },
+  container: {
+    flex: 1,
+    width: '100%', // 명시적으로 너비 부여
+  },
+  stepsWrapper: {
+    width: '100%',
+    alignItems: 'center', // 내부 OnboardingSteps를 위해 센터링 유지
+    paddingTop: 50,
+    paddingHorizontal: 24,
+    marginBottom: 8,
+    zIndex: 10,
+  },
+});
